refactor(base-gen): extract BatchProgressMixin (Tier-1b Task 7)
Move the batch-with-progress driver (_run_batch_with_progress) plus the
run_rerun public template method and its overridable rerun hooks
(_get_rerun_user_ids, _build_rerun_request_params, _create_run_request_for_item,
_create_rerun_response, _get_generated_count, _pre_process_rerun) verbatim into
base_generation/_batch_progress.py:BatchProgressMixin[TRequest], composed into
BaseGenerationService ahead of the existing mixins. None of the moved methods is
@abstractmethod; __abstractmethods__ frozenset stays 8. Annotation-only stub for
the base-owned _is_batch_mode field (sole writer); type-only stubs for base-owned
run/_create_state_manager/_get_base_service_name resolved via MRO. Behavior-preserving;
no public-surface, subclass-contract, or import-path change.
