+++ reflexio/server/services/base_generation/__init__.py
@@ -7,6 +7,9 @@ ABC skeleton (all abstract methods), ``__init__`` (per-run field inits), the ``r
 FIFO drain, ``_run_generation`` billing emit-ordering, and the lock/queue helpers.
 """
 
+from reflexio.server.services.base_generation._batch_progress import (
+    BatchProgressMixin,
+)
 from reflexio.server.services.base_generation._config_filter import (
     ConfigFilterMixin,
 )
@@ -25,6 +28,7 @@ from reflexio.server.services.base_generation._usage_billing import (
 )
 
 __all__ = [
+    "BatchProgressMixin",
     "ConfigFilterMixin",
     "ExtractionRunLifecycleMixin",
     "ShouldRunPrecheckMixin",
